Use uint64_t instead of ino_t
POSIX defines ino_t to be of an unsigned integer type, and searching
around the net didn't tell me of any definitions conflicting that. So
every ino_t can be represented in an uint64_t. (Assuming that is the
largest integer type in use for an inode number, but I'm sure that
assumption will hold for a while)

(dev_t, on the other hand, is a bit messier. Still figuring out what to
do with that.)

+++ src/dir_scan.c
@@ -49,7 +49,7 @@ static dev_t curdev;   /* current device we're scanning on */
 /* Populates the struct dir item with information from the stat struct. Sets
  * everything necessary for output_dir.item() except FF_ERR and FF_EXL. */
 static void stat_to_dir(struct dir *d, struct stat *fs) {
-  d->ino = fs->st_ino;
+  d->ino = (uint64_t)fs->st_ino;
   d->dev = fs->st_dev;
 
   if(S_ISREG(fs->st_mode))
